etl/requirements.txt
tushare
akshare
psycopg[binary]
pandas
python-dotenv
ta
